+++ app/src/main/AndroidManifest.xml
@@ -7,8 +7,11 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-        <activity android:name=".Activity.MainActivity">
+        android:theme="@android:style/Theme.Black.NoTitleBar"
+        >
+        <activity
+            android:name=".Activity.SplashActivity"
+            >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
