+++ app/src/main/java/zz/demoapp/Activity/DashBoardActivity.java
@@ -1,7 +1,7 @@
 package zz.demoapp.Activity;
 
+import android.app.Activity;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 
 import zz.demoapp.R;
 
@@ -9,7 +9,7 @@ import zz.demoapp.R;
  * Created by ankikumar on 5/28/2017.
  */
 
-public class DashBoardActivity extends AppCompatActivity{
+public class DashBoardActivity extends Activity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
