+++ app/src/main/res/anim/fade.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <alpha
+        android:fromAlpha="0"
+        android:toAlpha="1"
+        android:duration="2000"
+    ></alpha>
+    <alpha
+        android:startOffset="2000"
+        android:fromAlpha="1"
+        android:toAlpha="0"
+        android:duration="2000"
+        ></alpha>
+</set>
