+++ app/src/main/res/anim/move.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <translate
+        android:fromXDelta="-100"
+        android:toXDelta="0"
+        android:duration="2000"
+
+    ></translate>
+</set>
