+++ app/src/main/res/anim/scale.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <scale
+        android:fromXScale="0.33"
+        android:toXScale="1"
+        android:fromYScale="0.33"
+        android:toYScale="1"
+        android:pivotX="50%"
+        android:pivotY="50%"
+        android:duration="3000"
+    ></scale>
+</set>
