+++ app/src/main/res/layout/activity_splash.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
@@ -8,15 +9,20 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="zz.demoapp.Activity.MainActivity"
-    android:background="@drawable/black_bg"
+    tools:context="zz.demoapp.Activity.SplashActivity"
+    android:background="@drawable/frame"
     >
 
     <TextView
-        android:text="FlashScreen"
+        android:text="Relax"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_centerVertical="true"
         android:layout_centerHorizontal="true"
+        android:textSize="40sp"
+        android:textColor="@android:color/holo_blue_bright"
+        android:padding="10sp"
+        android:fontFamily="sans-serif-condensed"
         android:id="@+id/textView2" />
+
 </RelativeLayout>
